refactor: extract touch reference frame helpers
